fix: file corruption when calling parallel saves
Fixed by having two flags, one for saves in process and one to enqueue a
new save. When calling the commit method one or multiple times during an
already running save process, the task is essentially enqueued. We can
achieve this with only one flag since the data object is synchronously
up-to-date, so we only need to know if the user invoked the commit
method while we were still saving.

Closes #3.